app/src/main/java/com/kodebloc/hospitalmanagementproject/util/InputValidator.java
package com.kodebloc.hospitalmanagementproject.util;

import java.util.regex.Pattern;

public class InputValidator {

    // Updated email pattern to handle common valid formats
    private static final Pattern EMAIL_PATTERN = Pattern.compile(
            "^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$");

    public static boolean isValidEmail(String email) {
        return email != null && !email.isEmpty() && EMAIL_PATTERN.matcher(email).matches();
    }

    public static boolean isValidPassword(String password) {
        return password != null && password.length() >= 6;
    }
}
